+++ packages/core-js-compat/src/data.mjs
@@ -22,10 +22,28 @@ export const data = {
     rhino: '1.8.0',
     safari: '12.1',
   },
+  'es.symbol.async-dispose': {
+    bun: '1.0.23',
+    chrome: '127',
+    deno: '1.38',
+    firefox: '135',
+    // Node 20.4.0 add `Symbol.asyncDispose`, but with incorrect descriptor
+    // https://github.com/nodejs/node/issues/48699
+    node: '20.5.0',
+  },
   'es.symbol.async-iterator': {
     chrome: '63',
     firefox: '55',
     safari: '12.0',
+  },
+  'es.symbol.dispose': {
+    bun: '1.0.23',
+    chrome: '125',
+    deno: '1.38',
+    firefox: '135',
+    // Node 20.4.0 add `Symbol.dispose`, but with incorrect descriptor
+    // https://github.com/nodejs/node/issues/48699
+    node: '20.5.0',
   },
   'es.symbol.for': {
     chrome: '41',
@@ -139,6 +157,15 @@ export const data = {
     rhino: '1.8.0',
     safari: '15.0',
   },
+  'es.error.is-error': {
+    // early WebKit implementation bug
+    // https://github.com/oven-sh/bun/issues/15821
+    // bun: '1.1.39',
+    chrome: '134',
+    firefox: '138',
+    // https://github.com/nodejs/node/issues/58134
+    node: false,
+  },
   'es.error.to-string': {
     chrome: '33',
     firefox: '11',
@@ -164,6 +191,16 @@ export const data = {
     'react-native': '0.72',
     rhino: '1.8.0',
     safari: '15.0',
+  },
+  'es.suppressed-error.constructor': {
+    // Bun ~ 1.0.33 issues
+    // https://github.com/oven-sh/bun/issues/9282
+    // https://github.com/oven-sh/bun/issues/9283
+    bun: '1.2.15', // '1.0.23',
+    // reverted in https://issues.chromium.org/issues/42203506#comment25
+    // disabled again in 135 and re-enabled in 136
+    chrome: '136', // '134', // '133',
+    deno: '2.2.10',
   },
   'es.array.at': {
     chrome: '92',
@@ -448,7 +485,7 @@ export const data = {
     chrome: '110',
     deno: '1.27',
     // Incorrect exception thrown when index coercion fails
-    // firefox: '115',
+    firefox: '140', // '115',
     hermes: '0.13',
     'react-native': '0.74',
     rhino: '1.8.0',
@@ -590,6 +627,12 @@ export const data = {
     rhino: '1.7.13',
     safari: '3.1',
   },
+  'es.disposable-stack.constructor': {
+    // reverted in https://issues.chromium.org/issues/42203506#comment25
+    // disabled again in 135 and re-enabled in 136
+    chrome: '136', // '134', // '133',
+    deno: '2.2.10',
+  },
   'es.escape': {
     chrome: '1',
     firefox: '1',
@@ -638,6 +681,13 @@ export const data = {
     firefox: '131',
     safari: '18.4',
   },
+  'es.iterator.dispose': {
+    // reverted in https://issues.chromium.org/issues/42203506#comment25
+    // disabled again in 135 and re-enabled in 136
+    chrome: '136', // '134', // '133',
+    deno: '2.2.10',
+    firefox: '135',
+  },
   'es.iterator.drop': {
     // with changes related to the new iteration closing approach on early error
     // https://github.com/tc39/ecma262/pull/3467
@@ -645,7 +695,7 @@ export const data = {
     bun: '1.2.11', // '1.2.4', // '1.1.31',
     chrome: '135', // '122',
     deno: '2.2.5', // '1.38.1',
-    // firefox: '131',
+    firefox: '141', // '131',
     // safari: '18.4',
   },
   'es.iterator.every': {
@@ -654,7 +704,7 @@ export const data = {
     bun: '1.2.4', // '1.1.31',
     chrome: '135', // '122',
     deno: '2.2.5', // '1.38.1',
-    // firefox: '131',
+    firefox: '141', // '131',
     // safari: '18.4',
   },
   'es.iterator.filter': {
@@ -663,7 +713,7 @@ export const data = {
     bun: '1.2.4', // '1.1.31',
     chrome: '135', // '122',
     deno: '2.2.5', // '1.38.1',
-    // firefox: '131',
+    firefox: '141', // '131',
     // safari: '18.4',
   },
   'es.iterator.find': {
@@ -672,7 +722,7 @@ export const data = {
     bun: '1.2.4', // '1.1.31',
     chrome: '135', // '122',
     deno: '2.2.5', // '1.38.1',
-    // firefox: '131',
+    firefox: '141', // '131',
     // safari: '18.4',
   },
   'es.iterator.flat-map': {
@@ -681,7 +731,7 @@ export const data = {
     bun: '1.2.4', // '1.1.31',
     chrome: '135', // '122',
     deno: '2.2.5', // '1.38.1',
-    // firefox: '131',
+    firefox: '141', // '131',
     // safari: '18.4',
   },
   'es.iterator.for-each': {
@@ -690,15 +740,17 @@ export const data = {
     bun: '1.2.4', // '1.1.31',
     chrome: '135', // '122',
     deno: '2.2.5', // '1.38.1',
-    // firefox: '131',
+    firefox: '141', // '131',
     // safari: '18.4',
   },
   'es.iterator.from': {
-    bun: '1.1.31',
+    // Because of a bug in wrapper validation https://bugs.webkit.org/show_bug.cgi?id=288714
+    bun: '1.2.5',  // '1.1.31',
     chrome: '122',
     deno: '1.38.1',
     firefox: '131',
-    safari: '18.4',
+    // Because of a bug in wrapper validation https://bugs.webkit.org/show_bug.cgi?id=288714
+    // safari: '18.4',
   },
   'es.iterator.map': {
     // with changes related to the new iteration closing approach on early error
@@ -706,7 +758,7 @@ export const data = {
     bun: '1.2.4', // '1.1.31',
     chrome: '135', // '122',
     deno: '2.2.5', // '1.38.1',
-    // firefox: '131',
+    firefox: '141', // '131',
     // safari: '18.4',
   },
   'es.iterator.reduce': {
@@ -717,7 +769,7 @@ export const data = {
     bun: '1.2.11', // '1.2.4', // '1.1.31',
     chrome: '135', // '122',
     deno: '2.2.5', // '1.38.1',
-    // firefox: '131',
+    firefox: '141', // '131',
     // safari: '18.4',
   },
   'es.iterator.some': {
@@ -726,7 +778,7 @@ export const data = {
     bun: '1.2.4', // '1.1.31',
     chrome: '135', // '122',
     deno: '2.2.5', // '1.38.1',
-    // firefox: '131',
+    firefox: '141', // '131',
     // safari: '18.4',
   },
   'es.iterator.take': {
@@ -736,7 +788,7 @@ export const data = {
     bun: '1.2.11', // '1.2.4', // '1.1.31',
     chrome: '135', // '122',
     deno: '2.2.5', // '1.38.1',
-    // firefox: '131',
+    firefox: '141', // '131',
     // safari: '18.4',
   },
   'es.iterator.to-array': {
@@ -1389,6 +1441,22 @@ export const data = {
     chrome: '119',
     firefox: '121',
     safari: '17.4',
+  },
+  'es.array.from-async': { // <- `Array#values` and `Promise` dependencies should be loaded before
+    // https://bugs.webkit.org/show_bug.cgi?id=271703
+    bun: '1.1.2', // '0.3.0',
+    chrome: '121',
+    deno: '1.38',
+    firefox: '115',
+    // https://bugs.webkit.org/show_bug.cgi?id=271703
+    safari: '18.0', // '16.4',
+  },
+  'es.async-disposable-stack.constructor': { // `Promise` dependency should be loaded before
+    // added in 133, reverted in 134, https://issues.chromium.org/issues/42203506#comment25
+    // https://github.com/tc39/proposal-explicit-resource-management/issues/256, fixed in early 135
+    chrome: '136',
+  },
+  'es.async-iterator.async-dispose': { // `Promise` dependency should be loaded before
   },
   'es.reflect.apply': {
     chrome: '49',
@@ -2210,12 +2278,14 @@ export const data = {
     safari: '10.0',
   },
   'es.typed-array.with': {
-    bun: '0.1.9',
+    // It should truncate a negative fractional index to zero, but instead throws an error
+    // bun: '0.1.9',
     chrome: '110',
     deno: '1.27',
     firefox: '115',
     rhino: '1.8.0',
-    safari: '16.4',
+    // It should truncate a negative fractional index to zero, but instead throws an error
+    // safari: '16.4',
   },
   'es.unescape': {
     chrome: '1',
@@ -2251,25 +2321,10 @@ export const data = {
   },
   // TODO: Remove from `core-js@4`
   'esnext.aggregate-error': null,
-  'esnext.suppressed-error.constructor': {
-    // Bun ~ 1.0.33 issues
-    // https://github.com/oven-sh/bun/issues/9282
-    // https://github.com/oven-sh/bun/issues/9283
-    // bun: '1.0.23',
-    // reverted in https://issues.chromium.org/issues/42203506#comment25
-    // disabled again in 135 and re-enabled in 136
-    chrome: '136', // '134', // '133',
-    deno: '2.2.10',
-  },
-  'esnext.array.from-async': {
-    // https://bugs.webkit.org/show_bug.cgi?id=271703
-    bun: '1.1.2', // '0.3.0',
-    chrome: '121',
-    deno: '1.38',
-    firefox: '115',
-    // https://bugs.webkit.org/show_bug.cgi?id=271703
-    safari: '18.0', // '16.4',
-  },
+  // TODO: Remove from `core-js@4`
+  'esnext.suppressed-error.constructor': null,
+  // TODO: Remove from `core-js@4`
+  'esnext.array.from-async': null,
   // TODO: Remove from `core-js@4`
   'esnext.array.at': null,
   // TODO: Remove from `core-js@4`
@@ -2325,19 +2380,15 @@ export const data = {
   'esnext.array-buffer.transfer': null,
   // TODO: Remove from `core-js@4`
   'esnext.array-buffer.transfer-to-fixed-length': null,
-  'esnext.async-disposable-stack.constructor': {
-    // added in 133, reverted in 134, https://issues.chromium.org/issues/42203506#comment25
-    // https://github.com/tc39/proposal-explicit-resource-management/issues/256, fixed in early 135
-    chrome: '136',
-    deno: false,
-  },
+  // TODO: Remove from `core-js@4`
+  'esnext.async-disposable-stack.constructor': null,
   'esnext.async-iterator.constructor': {
   },
   // TODO: Remove from `core-js@4`
   'esnext.async-iterator.as-indexed-pairs': {
   },
-  'esnext.async-iterator.async-dispose': {
-  },
+  // TODO: Remove from `core-js@4`
+  'esnext.async-iterator.async-dispose': null,
   'esnext.async-iterator.drop': {
   },
   'esnext.async-iterator.every': {
@@ -2379,21 +2430,10 @@ export const data = {
   'esnext.data-view.set-float16': null,
   'esnext.data-view.set-uint8-clamped': {
   },
-  'esnext.disposable-stack.constructor': {
-    // reverted in https://issues.chromium.org/issues/42203506#comment25
-    // disabled again in 135 and re-enabled in 136
-    chrome: '136', // '134', // '133',
-    deno: '2.2.10',
-  },
-  'esnext.error.is-error': {
-    // early WebKit implementation bug
-    // https://github.com/oven-sh/bun/issues/15821
-    // bun: '1.1.39',
-    chrome: '134',
-    firefox: '138',
-    // https://github.com/nodejs/node/issues/58134
-    node: false,
-  },
+  // TODO: Remove from `core-js@4`
+  'esnext.disposable-stack.constructor': null,
+  // TODO: Remove from `core-js@4`
+  'esnext.error.is-error': null,
   'esnext.function.demethodize': {
   },
   'esnext.function.is-callable': {
@@ -2414,13 +2454,8 @@ export const data = {
   },
   'esnext.iterator.concat': {
   },
-  'esnext.iterator.dispose': {
-    // reverted in https://issues.chromium.org/issues/42203506#comment25
-    // disabled again in 135 and re-enabled in 136
-    chrome: '136', // '134', // '133',
-    deno: '2.2.10',
-    firefox: '135',
-  },
+  // TODO: Remove from `core-js@4`
+  'esnext.iterator.dispose': null,
   // TODO: Remove from `core-js@4`
   'esnext.iterator.drop': null,
   // TODO: Remove from `core-js@4`
@@ -2450,6 +2485,10 @@ export const data = {
   // TODO: Remove from `core-js@4`
   'esnext.iterator.to-array': null,
   'esnext.iterator.to-async': {
+  },
+  'esnext.iterator.zip': {
+  },
+  'esnext.iterator.zip-keyed': {
   },
   'esnext.json.is-raw-json': {
     bun: '1.1.43',
@@ -2519,6 +2558,7 @@ export const data = {
   // TODO: Remove from `core-js@4`
   'esnext.map.upsert': {
   },
+  // TODO: Remove from `core-js@4`
   'esnext.math.clamp': {
   },
   'esnext.math.deg-per-rad': {
@@ -2554,6 +2594,8 @@ export const data = {
   },
   // TODO: Remove from `core-js@4`
   'esnext.math.umulh': {
+  },
+  'esnext.number.clamp': {
   },
   'esnext.number.from-string': {
   },
@@ -2693,26 +2735,12 @@ export const data = {
   'esnext.string.replace-all': null,
   // TODO: Remove from `core-js@4`
   'esnext.string.to-well-formed': null,
-  'esnext.symbol.async-dispose': {
-    bun: '1.0.23',
-    chrome: '127',
-    deno: '1.38',
-    firefox: '135',
-    // Node 20.4.0 add `Symbol.asyncDispose`, but with incorrect descriptor
-    // https://github.com/nodejs/node/issues/48699
-    node: '20.5.0',
-  },
+  // TODO: Remove from `core-js@4`
+  'esnext.symbol.async-dispose': null,
   'esnext.symbol.custom-matcher': {
   },
-  'esnext.symbol.dispose': {
-    bun: '1.0.23',
-    chrome: '125',
-    deno: '1.38',
-    firefox: '135',
-    // Node 20.4.0 add `Symbol.dispose`, but with incorrect descriptor
-    // https://github.com/nodejs/node/issues/48699
-    node: '20.5.0',
-  },
+  // TODO: Remove from `core-js@4`
+  'esnext.symbol.dispose': null,
   'esnext.symbol.is-registered-symbol': {
   },
   // TODO: Remove from `core-js@4`
@@ -3064,9 +3092,11 @@ export const renamed = new Map([
   ['es.set', 'es.set.constructor'],
   ['es.weak-map', 'es.weak-map.constructor'],
   ['es.weak-set', 'es.weak-set.constructor'],
+  ['esnext.aggregate-error', 'es.aggregate-error'],
   ['esnext.array.at', 'es.array.at'],
   ['esnext.array.find-last', 'es.array.find-last'],
   ['esnext.array.find-last-index', 'es.array.find-last-index'],
+  ['esnext.array.from-async', 'es.array.from-async'],
   ['esnext.array.to-reversed', 'es.array.to-reversed'],
   ['esnext.array.to-sorted', 'es.array.to-sorted'],
   ['esnext.array.to-spliced', 'es.array.to-spliced'],
@@ -3074,11 +3104,15 @@ export const renamed = new Map([
   ['esnext.array-buffer.detached', 'es.array-buffer.detached'],
   ['esnext.array-buffer.transfer', 'es.array-buffer.transfer'],
   ['esnext.array-buffer.transfer-to-fixed-length', 'es.array-buffer.transfer-to-fixed-length'],
-  ['esnext.aggregate-error', 'es.aggregate-error'],
+  ['esnext.async-disposable-stack.constructor', 'es.async-disposable-stack.constructor'],
+  ['esnext.async-iterator.async-dispose', 'es.async-iterator.async-dispose'],
   ['esnext.data-view.get-float16', 'es.data-view.get-float16'],
   ['esnext.data-view.set-float16', 'es.data-view.set-float16'],
+  ['esnext.disposable-stack.constructor', 'es.disposable-stack.constructor'],
+  ['esnext.error.is-error', 'es.error.is-error'],
   ['esnext.global-this', 'es.global-this'],
   ['esnext.iterator.constructor', 'es.iterator.constructor'],
+  ['esnext.iterator.dispose', 'es.iterator.dispose'],
   ['esnext.iterator.drop', 'es.iterator.drop'],
   ['esnext.iterator.every', 'es.iterator.every'],
   ['esnext.iterator.filter', 'es.iterator.filter'],
@@ -3111,7 +3145,10 @@ export const renamed = new Map([
   ['esnext.string.match-all', 'es.string.match-all'],
   ['esnext.string.replace-all', 'es.string.replace-all'],
   ['esnext.string.to-well-formed', 'es.string.to-well-formed'],
+  ['esnext.suppressed-error.constructor', 'es.suppressed-error.constructor'],
   ['esnext.typed-array.at', 'es.typed-array.at'],
+  ['esnext.symbol.async-dispose', 'es.symbol.async-dispose'],
+  ['esnext.symbol.dispose', 'es.symbol.dispose'],
   ['esnext.typed-array.find-last', 'es.typed-array.find-last'],
   ['esnext.typed-array.find-last-index', 'es.typed-array.find-last-index'],
   ['esnext.typed-array.to-reversed', 'es.typed-array.to-reversed'],