+++ tests/compat/tests.js
@@ -374,8 +374,16 @@ GLOBAL.tests = {
     // eslint-disable-next-line symbol-description -- required for testing
     return Symbol('description detection').description === 'description detection' && Symbol().description === undefined;
   },
+  'es.symbol.async-dispose': function () {
+    var descriptor = Object.getOwnPropertyDescriptor(Symbol, 'asyncDispose');
+    return descriptor.value && !descriptor.enumerable && !descriptor.configurable && !descriptor.writable;
+  },
   'es.symbol.async-iterator': function () {
     return Symbol.asyncIterator;
+  },
+  'es.symbol.dispose': function () {
+    var descriptor = Object.getOwnPropertyDescriptor(Symbol, 'dispose');
+    return descriptor.value && !descriptor.enumerable && !descriptor.configurable && !descriptor.writable;
   },
   'es.symbol.for': SYMBOL_REGISTRY,
   'es.symbol.has-instance': [SYMBOLS_SUPPORT, function () {
@@ -419,6 +427,12 @@ GLOBAL.tests = {
     return new Error('e', { cause: 7 }).cause === 7
       && !('cause' in Error.prototype);
   },
+  'es.error.is-error': function () {
+    return PROTOTYPE_SETTING_AVAILABLE &&
+      (typeof DOMException != 'function' || Error.isError(new DOMException('DOMException'))) &&
+      Error.isError(new Error('Error', { cause: function () { /* empty */ } })) &&
+      !Error.isError(Object.create(Error.prototype));
+  },
   'es.error.to-string': function () {
     if (DESCRIPTORS_SUPPORT) {
       // Chrome 32- incorrectly call accessor
@@ -438,6 +452,11 @@ GLOBAL.tests = {
   'es.aggregate-error.cause': function () {
     return new AggregateError([1], 'e', { cause: 7 }).cause === 7
       && !('cause' in AggregateError.prototype);
+  },
+  'es.suppressed-error.constructor': function () {
+    return typeof SuppressedError == 'function'
+      && SuppressedError.length === 3
+      && new SuppressedError(1, 2, 3, { cause: 4 }).cause !== 4;
   },
   'es.array.at': function () {
     return [].at;
@@ -750,6 +769,9 @@ GLOBAL.tests = {
   'es.date.to-string': function () {
     return new Date(NaN).toString() === 'Invalid Date';
   },
+  'es.disposable-stack.constructor': function () {
+    return typeof DisposableStack == 'function';
+  },
   'es.escape': function () {
     return escape;
   },
@@ -775,6 +797,9 @@ GLOBAL.tests = {
         && Iterator.prototype === Object.getPrototypeOf(Object.getPrototypeOf([].values()));
     }
   },
+  'es.iterator.dispose': function () {
+    return [].keys()[Symbol.dispose];
+  },
   'es.iterator.drop': [
     iteratorHelperThrowsErrorOnInvalidIterator('drop', 0),
     checkIteratorClosingOnEarlyError('drop', RangeError)
@@ -791,7 +816,8 @@ GLOBAL.tests = {
   ],
   'es.iterator.for-each': checkIteratorClosingOnEarlyError('forEach', TypeError),
   'es.iterator.from': function () {
-    return Iterator.from;
+    Iterator.from({ 'return': null })['return']();
+    return true;
   },
   'es.iterator.map': [
     iteratorHelperThrowsErrorOnInvalidIterator('map', function () { /* empty */ }),
@@ -909,6 +935,7 @@ GLOBAL.tests = {
     return Math.trunc;
   },
   'es.number.constructor': function () {
+    // eslint-disable-next-line math/no-static-nan-calculations -- feature detection
     return Number(' 0o1') && Number('0b1') && !Number('+0x1');
   },
   'es.number.epsilon': function () {
@@ -1126,6 +1153,24 @@ GLOBAL.tests = {
   'es.promise.with-resolvers': [PROMISES_SUPPORT, function () {
     return Promise.withResolvers;
   }],
+  'es.array.from-async': function () {
+    // https://bugs.webkit.org/show_bug.cgi?id=271703
+    var counter = 0;
+    Array.fromAsync.call(function () {
+      counter++;
+      return [];
+    }, { length: 0 });
+    return counter === 1;
+  },
+  'es.async-disposable-stack.constructor': function () {
+    // https://github.com/tc39/proposal-explicit-resource-management/issues/256
+    // can't be detected synchronously
+    if (V8_VERSION && V8_VERSION < 136) return;
+    return typeof AsyncDisposableStack == 'function';
+  },
+  'es.async-iterator.async-dispose': function () {
+    return AsyncIterator.prototype[Symbol.asyncDispose];
+  },
   'es.reflect.apply': function () {
     try {
       return Reflect.apply(function () {
@@ -1661,13 +1706,21 @@ GLOBAL.tests = {
   'es.typed-array.to-sorted': function () {
     return Int8Array.prototype.toSorted;
   },
-  'es.typed-array.with': function () {
+  'es.typed-array.with': [function () {
     try {
       new Int8Array(1)['with'](2, { valueOf: function () { throw 8; } });
     } catch (error) {
       return error === 8;
     }
-  },
+  }, function () {
+    // WebKit doesn't handle this correctly. It should truncate a negative fractional index to zero, but instead throws an error
+    // Copyright (C) 2025 André Bargull. All rights reserved.
+    // This code is governed by the BSD license found in the LICENSE file.
+    // https://github.com/tc39/test262/pull/4477/commits/bd47071722d914036280cdd795a6ac6046d1c6f9
+    var ta = new Int8Array(1);
+    var result = ta['with'](-0.5, 1);
+    return result[0] === 1;
+  }],
   'es.unescape': function () {
     return unescape;
   },
@@ -1710,20 +1763,6 @@ GLOBAL.tests = {
       && set.add({}) === set
       && set[Symbol.toStringTag];
   }],
-  'esnext.suppressed-error.constructor': function () {
-    return typeof SuppressedError == 'function'
-      && SuppressedError.length === 3
-      && new SuppressedError(1, 2, 3, { cause: 4 }).cause !== 4;
-  },
-  'esnext.array.from-async': function () {
-    // https://bugs.webkit.org/show_bug.cgi?id=271703
-    var counter = 0;
-    Array.fromAsync.call(function () {
-      counter++;
-      return [];
-    }, { length: 0 });
-    return counter === 1;
-  },
   'esnext.array.filter-reject': function () {
     return [].filterReject;
   },
@@ -1733,17 +1772,8 @@ GLOBAL.tests = {
   'esnext.array.unique-by': function () {
     return [].uniqueBy;
   },
-  'esnext.async-disposable-stack.constructor': function () {
-    // https://github.com/tc39/proposal-explicit-resource-management/issues/256
-    // can't be detected synchronously
-    if (V8_VERSION && V8_VERSION < 136) return;
-    return typeof AsyncDisposableStack == 'function';
-  },
   'esnext.async-iterator.constructor': function () {
     return typeof AsyncIterator == 'function';
-  },
-  'esnext.async-iterator.async-dispose': function () {
-    return AsyncIterator.prototype[Symbol.asyncDispose];
   },
   'esnext.async-iterator.drop': function () {
     return AsyncIterator.prototype.drop;
@@ -1793,15 +1823,6 @@ GLOBAL.tests = {
   'esnext.data-view.set-uint8-clamped': [ARRAY_BUFFER_SUPPORT, function () {
     return DataView.prototype.setUint8Clamped;
   }],
-  'esnext.disposable-stack.constructor': function () {
-    return typeof DisposableStack == 'function';
-  },
-  'esnext.error.is-error': function () {
-    return PROTOTYPE_SETTING_AVAILABLE &&
-      (typeof DOMException != 'function' || Error.isError(new DOMException('DOMException'))) &&
-      Error.isError(new Error('Error', { cause: function () { /* empty */ } })) &&
-      !Error.isError(Object.create(Error.prototype));
-  },
   'esnext.function.demethodize': function () {
     return Function.prototype.demethodize;
   },
@@ -1817,14 +1838,17 @@ GLOBAL.tests = {
   'esnext.iterator.concat': function () {
     return Iterator.concat;
   },
-  'esnext.iterator.dispose': function () {
-    return [].keys()[Symbol.dispose];
-  },
   'esnext.iterator.range': function () {
     return Iterator.range;
   },
   'esnext.iterator.to-async': function () {
     return Iterator.prototype.toAsync;
+  },
+  'esnext.iterator.zip': function () {
+    return Iterator.zip;
+  },
+  'esnext.iterator.zip-keyed': function () {
+    return Iterator.zipKeyed;
   },
   'esnext.json.is-raw-json': NATIVE_RAW_JSON,
   'esnext.json.parse': function () {
@@ -1890,9 +1914,6 @@ GLOBAL.tests = {
   'esnext.map.update': function () {
     return Map.prototype.update;
   },
-  'esnext.math.clamp': function () {
-    return Math.clamp;
-  },
   'esnext.math.deg-per-rad': function () {
     return Math.DEG_PER_RAD;
   },
@@ -1916,6 +1937,9 @@ GLOBAL.tests = {
   },
   'esnext.math.sum-precise': function () {
     return Math.sumPrecise;
+  },
+  'esnext.number.clamp': function () {
+    return Number.prototype.clamp;
   },
   'esnext.number.from-string': function () {
     return Number.fromString;
@@ -1962,16 +1986,8 @@ GLOBAL.tests = {
   'esnext.string.dedent': function () {
     return String.dedent;
   },
-  'esnext.symbol.async-dispose': function () {
-    var descriptor = Object.getOwnPropertyDescriptor(Symbol, 'asyncDispose');
-    return descriptor.value && !descriptor.enumerable && !descriptor.configurable && !descriptor.writable;
-  },
   'esnext.symbol.custom-matcher': function () {
     return Symbol.customMatcher;
-  },
-  'esnext.symbol.dispose': function () {
-    var descriptor = Object.getOwnPropertyDescriptor(Symbol, 'dispose');
-    return descriptor.value && !descriptor.enumerable && !descriptor.configurable && !descriptor.writable;
   },
   'esnext.symbol.is-registered-symbol': function () {
     return Symbol.isRegisteredSymbol;